Constants start with k.
CANCoder config include whether it is reversed.

Arm controller in radians and not degrees (setAngle).

getCANCoder func named getCANCoder and not can coder rotations.

At the arm reset, don't return none if the arm is at -90 and in addition to cancoder reset, move the arm to -90.

At isReset don't check if the angle of the arm is -90, instead check tolerance.

Fill Outtake constants